pyproject.toml
[project]
name = "team8-fastapi"
version = "0.1.0"
description = "A FastAPI project for team8" # 설명 추가
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.1",
    "bcrypt<4.1.0",
    "black>=25.1.0",
    "cryptography>=45.0.3",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.13",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[project.optional-dependencies] # dependency-groups 대신 standard setuptools convention 사용
dev = [
    "black>=25.1.0",
    "ruff>=0.11.13",
]

[build-system]
requires = ["setuptools>=61.0.0"] # setuptools 버전 명시
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]       # 현재 디렉토리에서 패키지 검색 시작
include = ["app*"]  # 'app'으로 시작하는 디렉토리만 패키지로 포함 (즉, app/)
# exclude = ["migrations*"] # 필요하다면 migrations를 명시적으로 제외할 수 있지만, include만으로도 충분함

#[tool.uv]
#python = "3.13" # uv가 사용할 파이썬 버전을 명시하고 싶다면 주석 해제하여 사용